requirements_test.txt
coverage==5.1
mock>=4.0.2
flake8>=3.7.9
tox>=3.15.0
codecov>=2.0.22
django-model-utils==4.5.1

# Additional test requirements go here
django-test-plus==2.2.4
